cleanup and fixed warnings

+++ crc32.h
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #ifndef __CRC32_H
 #define __CRC32_H
 
-unsigned long crc32_buf(const void *buf, long len);
-unsigned long crc32_file(const char *filename);
+uint32_t crc32_buf(const void *buf, size_t len);
+uint32_t crc32_file(const char *filename);
 
 #endif
